+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - bug fixes:
   - modal: `css` not properly applied if the property was previously empty and the modal had been opened at least once
+  - image: allow overriding css background-image when value is empty
 
 ## 0.45.7